+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- [#214](https://github.com/deviceinsight/kafkactl/pull/214) Add `--context` option to set command's context
 
 ## 5.3.0 - 2024-08-14
 ### Added